Refactor the name of the test helper function.
The SenTest Framework runs all methods that return void, take no
parameters and whose name starts with `test`.

The helper function passes 2/3 of these and I think that it is not
immediately clear that it is a helper function. Not part of the
Objective-C coding style.

This commit renames:
- testPath:relativeToDirectory:expectedResult:

to:
- checkPath:relativeToDirectory:againstExpectedResult: